Refactor video embed handling into dedicated providers
