+++ plugin-notation-jeux_V4/includes/Video/Providers/VideoEmbedProviderInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace JLG\Notation\Video\Providers;
+
+interface VideoEmbedProviderInterface {
+
+    public function get_slug(): string;
+
+    public function build_embed_src( string $url ): string;
+}
